+++ NEWS.md
@@ -22,6 +22,8 @@ Red Panda C++ Version 0.13.3
   - fix: indent lines displayed at wrong position, when there are folded lines
   - fix: if editor's active line color is disabled, caret's position may not be correct redrawn
   - fix: insert code snippets will crash, if current compiler set's include dir list is not empty and lib dir list is empty
+  - fix: search around option can't be disabled
+  - enhancement: show a confirm dialog when search/replace around
   
 Red Panda C++ Version 0.13.2
   - fix: "delete and exit" button in the environtment / folder option page doesn't work correctly